+++ main.cpp
@@ -31,7 +31,7 @@ int main(int argc, char *argv[]) {
 
   //  auto start = high_resolution_clock::now();
     std::string outFile = argv[3];
-    double J = std::stod(argv[2]);
+    float J = std::stod(argv[2]);
     std:: cout << L << " " << J << std::endl;
     MC_Interacting_SAW_XY mcxysaw(L, J, outFile);
     mcxysaw.run_simulation(J);
